requirements.txt
torch>=2.2.0
numpy>=1.24.0
sounddevice>=0.4.6
requests>=2.31.0
wave>=0.0.2 
snac>=0.1.0
gradio>=4.19.2
huggingface_hub>=0.20.3
llama-cpp-python>=0.2.56
